+++ examples/cantherm/species/C2H4/ethene.py
@@ -22,7 +22,7 @@ opticalIsomers = 1
 energy = {
     'CBS-QB3': GaussianLog('ethene.log'),
     'Klip_2': -78.42735579,
-    'CCSD(T)-F12/cc-pVTZ-F12': MoleProLog('ethene_f12.out'),
+    'CCSD(T)-F12/cc-pVTZ-F12': MolproLog('ethene_f12.out'),
 }
 
 geometry = GaussianLog('ethene.log')
