+++ rmgpy/cantherm/statmech.py
@@ -43,7 +43,7 @@ import rmgpy.constants as constants
 
 from rmgpy.cantherm.output import prettify
 from rmgpy.cantherm.gaussian import GaussianLog
-from rmgpy.cantherm.molepro import MoleProLog 
+from rmgpy.cantherm.molpro import MolproLog
 from rmgpy.cantherm.qchem import QchemLog 
 
 from rmgpy.species import TransitionState, Species
@@ -202,7 +202,7 @@ class StatMechJob:
             # File formats
             'GaussianLog': GaussianLog,
             'QchemLog': QchemLog,
-            'MoleProLog': MoleProLog,
+            'MolproLog': MolproLog,
             'ScanLog': ScanLog,
         }
     
@@ -264,7 +264,7 @@ class StatMechJob:
         elif isinstance(energy, QchemLog):
             energyLog = energy; E0 = None
             energyLog.path = os.path.join(directory, energyLog.path)
-        elif isinstance(energy, MoleProLog):
+        elif isinstance(energy, MolproLog):
             energyLog = energy; E0 = None
             energyLog.path = os.path.join(directory, energyLog.path)
         elif isinstance(energy, float):
